rest_api/requirements.txt
# requirements.txt

Flask==2.3.2
Flask-HTTPAuth==4.8.0
Flask-CORS==3.0.10
psycopg2-binary==2.9.7
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-SocketIO==5.3.4 